requirements.txt
# SAP B1 Analytics Pack Dependencies
# Python 3.11+

# Database connectivity
pyodbc==5.0.1
pymssql==2.2.11
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data processing
pandas==2.1.4
numpy==1.26.2

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Scheduling
APScheduler==3.10.4
schedule==1.2.1

# Logging
coloredlogs==15.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Data validation
pydantic==2.5.2
pandera==0.17.2

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3.post1

# CLI
click==8.1.7
rich==13.7.0

# Development tools
black==23.12.1
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

